# Speed up insertion and deletion of values in SortedList by batching and deferring to idle task