Fixes #4911

Test setup:

- 10000 scrollback filled
- Decoration stress test button
- ~10000 * 20 = ~200,000 decorations

Before:

Insert ~4s
Delete ~4s

![image](https://github.com/xtermjs/xterm.js/assets/2193314/a1723f54-0304-4d35-9d1b-c4317c55f3e1)

After:

Insert ~330ms
Delete ~93ms

There's still a decent amount of GC happening on insert, but a lot of it has to happen due to the variables being used in callbacks. It's reasonable when you consider it's 200000 decorations.

![image](https://github.com/xtermjs/xterm.js/assets/2193314/70acdd4b-91cb-4197-9a94-e3cb81ac1639)